Replace title component with heading component
The heading component does not support a margin_top option and therefore the spacing was inconsistent with the title component (as the title component adds a top margin). I did explore changing the heading component to accept a margin_top option however this would involve making changes to the component wrapper helper or shared helper, which could cause this task to balloon a bit. As noted by matthillco, however, the main element in smart answers is currently missing a ‘govuk-main-wrapper’ class, and adding this seems to add appropriate spacing with a few visual caveats (please see the PR description for screenshots). Happy to discuss.