Fix error with leap years when calculating holiday entitlement
Due to how the calculation is performed to consider one year, we
were attempting to call a non-existent date e.g. 29th Feb 2025

Added logic in place to ensure that when we try to do this, we roll
to the 1st of March. This looks wrong, but gives us a correct
calculation response as determined in the tests.