Add handling for extreme date range to Smart Answers
There are cases where the user can type in a date, and sometimes this
leads to the user entering just two numbers instead of a full date.

Further down the stack this lead to incorrect dates and days being
recorded - for instance, "23" would lead to 0023 in the final result.

This new validation checks for cases where the date is less than the
year 100, implying only two digits entered, and brings the actual
date in line with expectations - essentially adding 2000 years.

It is a little bit of a workaround, however doing this validation
here can cover many instances where this is a problem.

It was first observed in the VAT Payment Deadline calculator.