Use Content Blocks in StatePensionThroughPartner flow
This replaces the use of content in a YAML file with content blocks
rendered via their appropriate embed codes. As these return HTML, with
the currencies already formatted, we don’t need to use the
`format_money` helpers.

If we wanted to use these values in calculations, we’d have to look at
a way of “embedding” content blocks without HTML wrappers.