Add a ContentBlockDetector class
This uses static analysis of code to fetch all the files associated
with a flow, reading the content of those files, and checking if any
content blocks embed codes have been used in them. If they have, we
return the content blocks in question.

I’ve added some YARD-compliant comments to the code, so it’s clear as
to what’s going on, as well as made a bunch of methods that would
otherwise be considered private to make unit testing easier.

In the next commit, we’ll use these to create links when publishing
smart answers in Publishing API.