Add a ContentBlockDetector class
This fetches all the files associated with a flow, reads the content
of those files, and checks if any content blocks embed codes have
been used in them. If they have, we return the content blocks in
question. In the next commit, we’ll use these to create links when
publishing smart answers in Publishing API.