Add range upper bounds to self assessment penalties
It was reported that if you submit a filing date far in the future, the smart answers calculator
for tax assessment penalties takes a massive performance hit trying to calculate the related years.

This commit adds a sensible upper range (the start of the tax year two years hence from today) to
the calculator, and changes the relevant error message on validation error.

Note that this commit does not change the limits on planned pay date to avoid any potential issues
with valid long term payment agreements made with HMRC.